lots of refactoring to make the tests adapt to the new factory.

+++ src/main/java/com/booleanuk/core/Order.java
@@ -15,13 +15,11 @@ public class Order {
         return basket;
     }
 
-    public Integer getTotalCost() {
+    public double getTotalCost() {
         for (Product product : this.basket.getProducts()) {
             this.totalCost += product.getPrice();
 
         }
-        return totalCost.intValue();
+        return totalCost;
     }
-
-
 }
